requirements.txt
streamlit>=1.34.0
google-generativeai
elevenlabs
pypdf
